Finished draft of downsizing for paired end. Needs to be tested.
